Move Shift+Enter to a separate function
