+++ lib/index.js
@@ -1,4 +1,5 @@
 const onEnter         = require('./onEnter');
+const onShiftEnter    = require('./onShiftEnter');
 const onTab           = require('./onTab');
 const onShiftTab      = require('./onShiftTab');
 const onBackspace     = require('./onBackspace');
@@ -47,6 +48,11 @@ function EditCode(opts) {
         // User is pressing Tab
         else if (data.key == KEY_TAB) {
             return onTab(...args);
+        }
+
+        // User is pressing Shift+Enter
+        else if (data.key == KEY_ENTER && data.isShift && opts.shiftEnterBlockType) {
+            return onShiftEnter(...args);
         }
 
         // User is pressing Enter
