+++ lib/onEnter.js
@@ -11,17 +11,6 @@ function onEnter(event, data, state, opts) {
 
     event.preventDefault();
 
-    if (data.isShift && opts.shiftEnterBlockType) {
-        // Exit the code block
-        const transform = state.transform();
-        transform.insertBlock({ type: opts.shiftEnterBlockType });
-
-        const inserted = transform.state.startBlock;
-        return transform.unwrapNodeByKey(inserted.key).apply();
-    }
-
-    event.preventDefault();
-
     const { startBlock } = state;
     const currentLineText = startBlock.text;
     const indent = getIndent(currentLineText, '');
